Moved NodeItem class to node_parser.py as more relevant. Code improvements.
