+++ src/hdusd/mx_nodes/nodes/__init__.py
@@ -51,9 +51,15 @@ def unregister():
     unregister_sockets()
 
 
-def get_node_def_cls(node_name, nd_type, prefix='STD'):
-    name = f"MxNodeDef_{prefix}_ND_{node_name}_{nd_type}"
-    return next(cls for cls in node_def_classes if cls.__name__ == name)
+def get_node_def_cls(node_name, nd_type):
+    nd_name = f"ND_{node_name}_{nd_type}"
+    node_def_cls = next((cls for cls in node_def_classes if cls.mx_nodedef.getName() == nd_name),
+                        None)
+    if node_def_cls:
+        return node_def_cls
+
+    nd_name = f"ND_{node_name}"
+    return next(cls for cls in node_def_classes if cls.mx_nodedef.getName() == nd_name)
 
 
 def get_mx_node_cls(node_name, prefix='STD'):
